Plan for circle
I preliminarily write down the plan for  circle: LCPosition and CCPosition. I feel that a core issue is to construct two intersections of a ray with a circle. I believe that the process can be done via the discussion in advance. My problem is that I still feel mystified how to implement this process.